Extract Inventory data prepare to ShardingScalingPreparer
