+++ sharding-scaling/sharding-scaling-core/src/main/java/org/apache/shardingsphere/shardingscaling/core/preparer/ShardingScalingJobPreparer.java
@@ -18,11 +18,15 @@
 package org.apache.shardingsphere.shardingscaling.core.preparer;
 
 import org.apache.shardingsphere.shardingscaling.core.ShardingScalingJob;
+import org.apache.shardingsphere.shardingscaling.core.config.SyncConfiguration;
 import org.apache.shardingsphere.shardingscaling.core.controller.task.SyncTaskControlStatus;
 import org.apache.shardingsphere.shardingscaling.core.datasource.DataSourceManager;
 import org.apache.shardingsphere.shardingscaling.core.exception.DatasourceCheckFailedException;
 import org.apache.shardingsphere.shardingscaling.core.preparer.checker.DataSourceChecker;
 import org.apache.shardingsphere.shardingscaling.core.preparer.checker.DataSourceCheckerCheckerFactory;
+import org.apache.shardingsphere.shardingscaling.core.preparer.splitter.InventoryDataTaskSplitter;
+import org.apache.shardingsphere.shardingscaling.core.synctask.inventory.InventoryDataSyncTaskGroup;
+
 import lombok.extern.slf4j.Slf4j;
 
 /**
@@ -30,6 +34,8 @@ import lombok.extern.slf4j.Slf4j;
  */
 @Slf4j
 public final class ShardingScalingJobPreparer {
+    
+    private final InventoryDataTaskSplitter inventoryDataTaskSplitter = new InventoryDataTaskSplitter();
     
     /**
      * Do prepare work for sharding scaling job.
@@ -40,6 +46,7 @@ public final class ShardingScalingJobPreparer {
         String databaseType = shardingScalingJob.getSyncConfigurations().get(0).getDumperConfiguration().getDataSourceConfiguration().getDatabaseType().getName();
         try (DataSourceManager dataSourceManager = new DataSourceManager(shardingScalingJob.getSyncConfigurations())) {
             checkDatasources(databaseType, dataSourceManager);
+            splitTasks(shardingScalingJob, dataSourceManager);
         } catch (DatasourceCheckFailedException ex) {
             log.warn("Preparing sharding scaling job {} : {} failed", shardingScalingJob.getJobId(), shardingScalingJob.getJobName(), ex);
             shardingScalingJob.setStatus(SyncTaskControlStatus.PREPARING_FAILURE.name());
@@ -50,5 +57,11 @@ public final class ShardingScalingJobPreparer {
         DataSourceChecker dataSourceChecker = DataSourceCheckerCheckerFactory.newInstanceDataSourceChecker(databaseType);
         dataSourceChecker.checkConnection(dataSourceManager.getCachedDataSources().values());
         dataSourceChecker.checkPrivilege(dataSourceManager.getSourceDatasources().values());
+    }
+    
+    private void splitTasks(final ShardingScalingJob shardingScalingJob, final DataSourceManager dataSourceManager) {
+        for (SyncConfiguration each : shardingScalingJob.getSyncConfigurations()) {
+            shardingScalingJob.getInventoryDataTasks().add(new InventoryDataSyncTaskGroup(each, inventoryDataTaskSplitter.splitInventoryData(each, dataSourceManager)));
+        }
     }
 }
