+++ sharding-scaling/sharding-scaling-core/src/main/java/org/apache/shardingsphere/shardingscaling/core/synctask/SyncTaskFactory.java
@@ -17,6 +17,8 @@
 
 package org.apache.shardingsphere.shardingscaling.core.synctask;
 
+import java.util.Collection;
+
 import org.apache.shardingsphere.shardingscaling.core.config.SyncConfiguration;
 import org.apache.shardingsphere.shardingscaling.core.synctask.inventory.InventoryDataSyncTask;
 import org.apache.shardingsphere.shardingscaling.core.synctask.inventory.InventoryDataSyncTaskGroup;
@@ -32,10 +34,10 @@ public interface SyncTaskFactory {
      * Create inventory data sync task group.
      *
      * @param syncConfiguration sync configuration
-     * @param dataSourceManager  data source factory
+     * @param inventoryDataSyncTasks  inventory data sync tasks
      * @return inventory data sync task group
      */
-    InventoryDataSyncTaskGroup createInventoryDataSyncTaskGroup(SyncConfiguration syncConfiguration, DataSourceManager dataSourceManager);
+    InventoryDataSyncTaskGroup createInventoryDataSyncTaskGroup(SyncConfiguration syncConfiguration, Collection<SyncTask> inventoryDataSyncTasks);
     
     /**
      * Create inventory data sync task.
